Refactor: Move actions-hub and preview-site folder under power-pages
