+++ .vscode/settings.json
@@ -1,10 +1,11 @@
 {
-    "cSpell.words": [
-      "lcid",
-      "nupkg",
-      "powerapps",
-      "powerpages",
-      "powerplatform"
-    ],
-    "sarif-viewer.connectToGithubCodeScanning": "off"
+  "cSpell.words": [
+    "dataverse",
+    "lcid",
+    "nupkg",
+    "powerapps",
+    "powerpages",
+    "powerplatform"
+  ],
+  "sarif-viewer.connectToGithubCodeScanning": "off"
 }
