+++ src/client/test/Integration/client/power-pages/actions-hub/tree-items/InactiveGroupTreeItem.test.ts
@@ -5,8 +5,8 @@
 
 import * as vscode from "vscode";
 import { expect } from "chai";
-import { ActionsHubTreeItem } from "../../../../../actions-hub/tree-items/ActionsHubTreeItem";
-import { InactiveGroupTreeItem } from "../../../../../actions-hub/tree-items/InactiveGroupTreeItem";
+import { InactiveGroupTreeItem } from "../../../../../../power-pages/actions-hub/tree-items/InactiveGroupTreeItem";
+import { ActionsHubTreeItem } from "../../../../../../power-pages/actions-hub/tree-items/ActionsHubTreeItem";
 
 describe('InactiveGroupTreeItem', () => {
     it('should be of type ActionsHubTreeItem', () => {
