+++ src/client/test/Integration/client/power-pages/actions-hub/tree-items/NoDataTreeItem.test.ts
@@ -5,8 +5,8 @@
 
 import * as vscode from "vscode";
 import { expect } from "chai";
-import { ActionsHubTreeItem } from "../../../../../actions-hub/tree-items/ActionsHubTreeItem";
-import { NoDataTreeItem } from "../../../../../actions-hub/tree-items/NoDataTreeItem";
+import { NoDataTreeItem } from "../../../../../../power-pages/actions-hub/tree-items/NoDataTreeItem";
+import { ActionsHubTreeItem } from "../../../../../../power-pages/actions-hub/tree-items/ActionsHubTreeItem";
 
 describe('NoDataTreeItem', () => {
     it('should be of type ActionsHubTreeItem', () => {
